fix(storage): extract validateDataRoot to dataRootValidator.ts — fix max-lines lint (t/3296)
fileIO.ts exceeded the 1500-line ESLint max-lines limit after the t/3296 A arm
addition. Extracts validateDataRoot() to storage/dataRootValidator.ts following
the existing ADR-007 cohesion-split pattern (debateStore, calibrationStore, etc.).
fileIO.ts re-exports it from the barrel; tests unchanged (import fileIO.js barrel).
fileIO.ts now 1486 lines.
